moving gamecontrol out of hexboard into core
